cmdline reorg add threed scripts
