+++ openst/cli.py
@@ -1149,6 +1149,181 @@ def cmd_run_pairwise_aligner(args):
     _run_pairwise_aligner(args)
 
 
+def get_from_3d_registration_parser():
+    """
+    Parse command-line arguments.
+
+    Returns:
+        argparse.Namespace: Parsed command-line arguments.
+    """
+    parser = argparse.ArgumentParser(
+        description="convert openst data for STIM (serial-section 3D registration); one file at a time",
+        allow_abbrev=False,
+        add_help=False,
+    )
+    parser.add_argument(
+        "--n5-dirs",
+        type=str,
+        nargs="+",
+        required=True,
+        help="Path to the n5 directories created by STIM for each of the samples. Expects more than one argument.",
+    )
+    parser.add_argument(
+        "--h5-files",
+        type=str,
+        nargs="+",
+        required=True,
+        help="""Path to the h5ad files used to create the STIM data.
+        Expects more than one argument, in the same order as --n5-dirs.""",
+    )
+    parser.add_argument(
+        "--images",
+        type=str,
+        nargs="+",
+        help="""Path to the file to load pairwise aligned images from.
+        In this case, it should have the same length as the --n5-dirs and --h5-files arguments.
+        If --images-in-adata is provided, it is the path inside the h5ad files where to load the images from.
+        In this case, there only one argument is provided.""",
+    )
+    parser.add_argument(
+        "--images-in-adata",
+        default=False,
+        action="store_true",
+        help="""When specified, the pairwise-aligned image is loaded from the adata,
+        at the internal path specified by '--images'""",
+    )
+    parser.add_argument(
+        "--output-h5",
+        type=str,
+        required=True,
+        help="""Path to output h5ad file that will be created.
+        When --separate-images is specified, it will not contain images.""",
+    )
+    parser.add_argument(
+        "--output-image",
+        type=str,
+        help="""Path to output h5ad file that will be created.
+        When --separate-images is specified, it will not contain images.""",
+    )
+    parser.add_argument(
+        "--save-images-separately",
+        default=False,
+        action="store_true",
+        help="When images are provided, these will be saved separately",
+    )
+    parser.add_argument(
+        "--rescale",
+        type=float,
+        default=1,
+        help="Rescales (multiples) the coordinates by --rescale units",
+    )
+    parser.add_argument(
+        "--downsample-iamge",
+        type=int,
+        default=1,
+        help="""How much to downwsample the resolution of the final aligned image volume
+        (recommended for very large images). Must be >= 1""",
+    )
+    parser.add_argument(
+        "--merge-output",
+        type=str,
+        help='how to merge tiles, can be "same", "unique", "first", "only"',
+        choices=["same", "unique", "first", "only"],
+        default="same",
+    )
+    parser.add_argument(
+        "--join-output",
+        type=str,
+        help='how to join tiles, can be "inner", "outer"',
+        choices=["inner", "outer"],
+        default="outer",
+    )
+    parser.add_argument(
+        "--max-image-pixels",
+        type=int,
+        default=933120000,
+        help="Upper bound for number of pixels in the images (prevents exception when opening very large images)",
+    )
+    return parser
+
+
+def setup_from_3d_registration_parser(parent_parser):
+    """setup_from_3d_registration_parser"""
+    parser = parent_parser.add_parser(
+        "from_3d_registration",
+        help="convert STIM-registered data into h5ad files",
+        parents=[get_from_3d_registration_parser()],
+    )
+    parser.set_defaults(func=cmd_run_from_3d_registration)
+
+    return parser
+
+def cmd_run_from_3d_registration(args):
+    from openst.threed.from_3d_registration import _run_from_3d_registration
+    _run_from_3d_registration(args)
+
+
+def get_to_3d_registration_parser():
+    """
+    Parse command-line arguments.
+
+    Returns:
+        argparse.Namespace: Parsed command-line arguments.
+    """
+    parser = argparse.ArgumentParser(
+        description="convert openst data for STIM (serial-section 3D registration); one file at a time",
+        allow_abbrev=False,
+        add_help=False,
+    )
+    parser.add_argument(
+        "--in-adata",
+        type=str,
+        required=True,
+        help="Path to the input adata file. Expects a file with raw counts.",
+    )
+    parser.add_argument(
+        "--output-dir",
+        type=int,
+        default=-1,
+        help="Path to the directory where the .genes and .locations csv files will be stored.",
+    )
+    parser.add_argument(
+        "--filter-umi-max",
+        type=int,
+        default=-1,
+        help="Preserve cells with at most < --filter-umi-max UMIs",
+    )
+    parser.add_argument(
+        "--rescale",
+        type=float,
+        default=1,
+        help="Rescales (multiples) the coordinates by --rescale units",
+    )
+    parser.add_argument(
+        "--genes",
+        type=str,
+        nargs="+",
+        default=None,
+        help="Exports the expression level for the specified genes",
+    )
+    return parser
+
+
+def setup_to_3d_registration_parser(parent_parser):
+    """setup_to_3d_registration_parser"""
+    parser = parent_parser.add_parser(
+        "to_3d_registration",
+        help="convert h5ad files to native STIM-supported format",
+        parents=[get_to_3d_registration_parser()],
+    )
+    parser.set_defaults(func=cmd_run_to_3d_registration)
+
+    return parser
+
+def cmd_run_to_3d_registration(args):
+    from openst.threed.to_3d_registration import _run_to_3d_registration
+    _run_to_3d_registration(args)
+
 def cmdline_args():
     parent_parser = argparse.ArgumentParser(
         allow_abbrev=False,
